backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic[email]==2.5.0
python-multipart==0.0.6
pytest==7.4.3
httpx==0.25.2
requests==2.31.0
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
authlib==1.3.0
cryptography
bcrypt
email-validator
firebase-admin
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
schedule==1.2.0
pywebpush==1.14.0

